OSS Friends list is out of date
