+++ packages/twenty-website/src/app/oss-friends/page.tsx
@@ -1,8 +1,6 @@
 import { Background } from '@/app/components/oss-friends/Background';
-import {
-  CardContainer,
-  OssData,
-} from '@/app/components/oss-friends/CardContainer';
+import { Card, OssData } from '@/app/components/oss-friends/Card';
+import { CardContainer } from '@/app/components/oss-friends/CardContainer';
 import { ContentContainer } from '@/app/components/oss-friends/ContentContainer';
 import { Header } from '@/app/components/oss-friends/Header';
 
@@ -16,20 +14,11 @@ export default async function OssFriends() {
       <Background />
       <ContentContainer>
         <Header />
-        <div
-          style={{
-            display: 'flex',
-            flexWrap: 'wrap',
-            gap: '32px',
-            justifyContent: 'center',
-            maxWidth: '1000px',
-            margin: '0 auto',
-          }}
-        >
+        <CardContainer>
           {listJson.data.map((data: OssData, index: number) => (
-            <CardContainer key={index} data={data} />
+            <Card key={index} data={data} />
           ))}
-        </div>
+        </CardContainer>
       </ContentContainer>
     </>
   );
